Added Docs, major refactor with KeynesianModel, and added Demos
Docs: It was stupid to develop this without some sort of reference.
This stupidity has been alleviated.

Demo: Mostly a playground for me to work with what I'm developing,
eventually it will be for real demos.

config.ru: simple rack application to server up static assets. This
is so I don't have to code relative paths.

KeynesianModel is now `Keynes.Model`