Has many relationship working. Added more documentation.
Still under heavy development. The has_many relationship has been added
and `seems` to work. I'll be adding tests in the morning, and starting
work on the `belongs_to` relationship next.

The documentation may be a bit wonky as its being refactored/updated,
rather constantly. Ideally, it'll be powered by Keynes itself before
too long.

There are basic examples of the models inside of the `demos` directory,
and the demo.html is currently what I'm using as my 'live' environment.